refactor: schedule 및 integration schedule 통합
