+++ backend/src/main/java/com/allog/dallog/domain/schedule/application/SubscribingSchedulesFinder.java
@@ -7,8 +7,8 @@ import com.allog.dallog.domain.auth.dto.response.OAuthAccessTokenResponse;
 import com.allog.dallog.domain.category.application.ExternalCategoryDetailService;
 import com.allog.dallog.domain.category.domain.ExternalCategoryDetail;
 import com.allog.dallog.domain.externalcalendar.application.ExternalCalendarClient;
-import com.allog.dallog.domain.integrationschedule.domain.IntegrationSchedule;
-import com.allog.dallog.domain.integrationschedule.domain.TypedSchedules;
+import com.allog.dallog.domain.schedule.domain.IntegrationSchedule;
+import com.allog.dallog.domain.schedule.domain.TypedSchedules;
 import com.allog.dallog.domain.schedule.dto.request.DateRangeRequest;
 import com.allog.dallog.domain.schedule.dto.response.MemberScheduleResponses;
 import com.allog.dallog.domain.subscription.domain.SubscriptionRepository;
