+++ backend/src/main/java/com/allog/dallog/domain/schedule/domain/ScheduleType.java
@@ -1,4 +1,4 @@
-package com.allog.dallog.domain.integrationschedule.domain;
+package com.allog.dallog.domain.schedule.domain;
 
 import java.util.Arrays;
 import java.util.function.Predicate;
