+++ backend/src/main/java/com/allog/dallog/domain/schedule/dto/response/PeriodResponse.java
@@ -1,6 +1,6 @@
 package com.allog.dallog.domain.schedule.dto.response;
 
-import com.allog.dallog.domain.integrationschedule.domain.Period;
+import com.allog.dallog.domain.schedule.domain.Period;
 import java.time.LocalDateTime;
 
 public class PeriodResponse {
